+++ previous-versions/assess/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.32.1] previous-versions/assess
 ### Releases
 - [PyPI v0.32.1] https://pypi.org/project/codat-assess/0.32.1 - previous-versions/assess
+
+## 2023-09-28 00:12:24
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.91.2 (2.131.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.32.2] previous-versions/assess
+### Releases
+- [PyPI v0.32.2] https://pypi.org/project/codat-assess/0.32.2 - previous-versions/assess